% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr-engine.R
\name{eng_wigo}
\alias{eng_wigo}
\title{An environment reporting engine for knitr}
\usage{
eng_wigo(options)
}
\arguments{
\item{options}{Chunk options provided by `knitr`}
}
\description{
This provides an engine for `knitr` which executes all underlying R code
but instead of outputting code results outputs a report about the environment
and changes to its states. This can be useful to determine what is going on
when an RMarkdown has unexpected behavior or when reading an RMarkdown for the
first time.
}
\details{
The engine can be called for all chunks by setting

```
library(wigo)
```

at the top of your RMarkown, or by putting `engine = wigo` within the header
of each individual R chunk.
}
