% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eng-diff.R
\name{eng_wigo_diff}
\alias{eng_wigo_diff}
\title{An environment diff reporting engine for knitr}
\usage{
eng_wigo_diff(options)
}
\arguments{
\item{options}{Chunk options provided by `knitr`}
}
\description{
This provides an engine for `knitr` which executes all underlying R code
but instead of outputting code results outputs a report about changes to
the environment state. This can be useful to determine what is going on
when an RMarkdown has unexpected behavior or when reading an RMarkdown for the
first time.
}
\details{
The engine can be called for all chunks by setting

```
library(wigo)
register_eng_wigo()
```
}
