% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_dimn}
\alias{get_dimn}
\title{Provides `length` for all object types except `data.frame`s and `nrow x ncol` for `data.frame`}
\usage{
get_dimn(x)
}
\description{
Provides `length` for all object types except `data.frame`s and `nrow x ncol` for `data.frame`
}
\keyword{internal}
