% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr-engine.R
\name{set_eng_wigo}
\alias{set_eng_wigo}
\title{Convenience function to expose wigo as a knitr engine}
\usage{
set_eng_wigo()
}
\value{
No return
}
\description{
This is a basic wrapper around `knitr::knit_engines$set` to expose `wigo`
as a valid and callable language engine for knitting an RMarkdown
}
\examples{
\dontrun{set_eng_wigo()}
}
