% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tbl_environ}
\alias{tbl_environ}
\title{Create `data.frame` representation of current environment}
\usage{
tbl_environ(environ, chunk_name)
}
\description{
Create `data.frame` representation of current environment
}
\keyword{internal}
