% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wigo-package.R
\docType{package}
\name{wigo-package}
\alias{wigo}
\alias{wigo-package}
\title{wigo: knitr engine to explain what is going on}
\description{
This knitr engine that outputs table representing state of
    knitr environment and key state changes. It is intended to help understand 
    and debug an RMarkdown script.
}
\author{
\strong{Maintainer}: Emily Riederer \email{emilyriederer@gmail.com}

}
\keyword{internal}
