% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eng-instances.R
\name{eng_wigo}
\alias{eng_wigo}
\alias{eng_wigo_diff}
\title{An environment reporting engine for knitr}
\usage{
eng_wigo(options)

eng_wigo_diff(options)
}
\arguments{
\item{options}{Chunk options provided by `knitr`}
}
\description{
This provides an engine for `knitr` which executes all underlying R code
but instead of outputting code results outputs a report about the environment
and changes to its states. This can be useful to determine what is going on
when an RMarkdown has unexpected behavior or when reading an RMarkdown for the
first time.
}
\details{
The engine can be called for all chunks by setting

```
library(wigo)
register_eng_wigo()
```

`wigo` and `wigo_diff` vary simply by which variables they chose to report.
`wigo` reports on the state of all variables in the environment, whereas
`wigo_diff` reports only on those that were changed (created, modified, or
removed) in the preceding chunk.
}
